@@ -14,20 +14,22 @@
 		"format": "prettier --ignore-path .gitignore --write --plugin-search-dir=. ."
 	},
 	"devDependencies": {
-		"@babel/core": "^7.17.9",
+		"@babel/core": "^7.18.5",
 		"@macfja/svelte-multi-adapter": "^1.0.2",
-		"@sveltejs/adapter-netlify": "^1.0.0-next.55",
-		"@sveltejs/adapter-node": "^1.0.0-next.73",
-		"@sveltejs/adapter-static": "^1.0.0-next.29",
+		"@sveltejs/adapter-netlify": "^1.0.0-next.65",
+		"@sveltejs/adapter-node": "^1.0.0-next.78",
+		"@sveltejs/adapter-static": "^1.0.0-next.34",
 		"@sveltejs/kit": "1.0.0-next.320",
 		"@types/babel-core": "^6.25.7",
+		"@types/dateformat": "^5.0.0",
 		"@types/js-cookie": "^3.0.2",
 		"@types/lodash-es": "^4.17.6",
 		"@typescript-eslint/eslint-plugin": "^4.33.0",
 		"@typescript-eslint/parser": "^4.33.0",
-		"autoprefixer": "^10.4.5",
+		"autoprefixer": "^10.4.4",
 		"babel-cli": "^6.26.0",
 		"cross-env": "^7.0.3",
+		"daisyui": "^2.17.0",
 		"dotenv": "^10.0.0",
 		"emailjs": "^3.8.1",
 		"eslint": "^7.32.0",
@@ -36,28 +38,31 @@
 		"object-sizeof": "^1.6.3",
 		"postcss": "^8.4.12",
 		"postcss-html": "^1.4.1",
+		"postcss-load-config": "^3.1.4",
 		"postcss-nested": "^5.0.6",
-		"postcss-scss": "^4.0.3",
+		"postcss-scss": "^4.0.4",
 		"prettier": "~2.2.1",
 		"prettier-plugin-svelte": "^2.7.0",
 		"redis": "^3.1.2",
-		"sass": "^1.50.1",
-		"svelte": "^3.47.0",
-		"svelte-check": "^2.7.0",
+		"sass": "^1.53.0",
+		"svelte": "^3.48.0",
+		"svelte-check": "^2.8.0",
 		"svelte-image": "^0.2.9",
-		"svelte-preprocess": "^4.10.6",
+		"svelte-preprocess": "^4.10.5",
 		"svelte-preprocess-import-assets": "^0.1.6",
 		"svelte-sequential-preprocessor": "^0.0.7",
 		"svelte-use-click-outside": "^1.0.0",
+		"tailwindcss": "^3.0.23",
 		"tslib": "^2.4.0",
-		"typescript": "^4.6.3",
-		"vite": "^2.9.5",
+		"typescript": "^4.7.4",
+		"vite": "^2.9.12",
 		"vite-imagetools": "^3.9.0"
 	},
 	"type": "module",
 	"dependencies": {
-		"firebase": "^9.6.11",
-		"functional-utilities": "^1.14.1",
+		"dateformat": "^5.0.3",
+		"firebase": "^9.8.4",
+		"functional-utilities": "^1.15.1",
 		"html2canvas": "^1.4.1",
 		"ibantools": "^4.1.5",
 		"js-cookie": "^3.0.1",
