@@ -1,10 +1,12 @@
 <script>
 	import Button from '../elements/interactive/buttons/button.svelte';
+	import SideBySide from '../layout/side_by_side.svelte';
 </script>
 
 <h2 class="gradient_text top_margin">Im Angebot inklusive</h2>
-<div class="side_alignment">
-	<div class="left">
+<SideBySide primary="left">
+	<span slot="left">
+		<h3 class="gradient_text">Sky Q</h3>
 		<p>
 			Mit unserem Sky Angebot können Sie Ihr Wunschprogramm über die gesamte Laufzeit von 12 Monaten
 			zu Neukunden-Preisen genießen. Der Sky Vertrag hat eine Laufzeit von 12 Monaten und ist nach 1
@@ -24,11 +26,61 @@
 			> an.
 		</p>
 		<div class="btn">
-			<Button text="Alle Infomationen zu Sky Q" route="/sky_q" />
+			<Button text="Alle Infomationen zu Sky Q" on_click="/sky_q" />
 		</div>
-	</div>
-	<img src="/images/visuals/skyq_header_4.png" alt="" />
-</div>
+	</span>
+	<span slot="right">
+		<img src="/images/visuals/skyq_header_4.png" alt="" />
+	</span>
+</SideBySide>
+<SideBySide primary="right">
+	<span slot="right">
+		<h3 class="gradient_text">Peacock für alle Sky Q Kunden</h3>
+		<ul>
+			<li>
+				<b>Exklusive Peacock Originals,</b> darunter die aktuellen Serien "Saved by the Bell", "Rutherford
+				Falls" und "Girls5eva"; und zusätzlich die kommenden Serien "The Girl in the Woods", "MacGruber",
+				"Joe vs. Carole", "Bel-Air" und "The Resort".
+			</li>
+			<li>
+				<b>Alle Staffeln der beliebtesten Drama und Comedy-Serien</b> wie "The Office", "Monk", "30 Rock",
+				"Psych" und "Will & Grace"; unverzichtbare Serien wie "Downton Abbey", "Dr. House", "Battlestar
+				Galactica", "Bates Motel" und "Heroes"; die beliebten Franchise Hits "Chicago Fire", "Chicago
+				P.D." und "Chicago Med"; und Reality-TV-Shows wie "Keeping Up With the Kardashians", "Below Deck"
+				und "Botched".
+			</li>
+			<li>
+				<b>Filme der letzten Jahre</b> wie "American Pie 2", "Back to the Future Part III", "Brüno",
+				"Bulletproof", "Children of Men", "Erin Brockovich", "The Game", "Gladiator", "The Green Mile",
+				"The Holiday", "Jason Bourne", "Identity Thief", "Lucy", "MacGruber", "Mamma Mia!", "Meet Joe
+				Black", "Notting Hill", "Pitch Perfect", "The Purge", "The Secret Life of Pets", "Shaun of the
+				Dead", "Snow White and the Huntsman", "Out of Sight", "Tower Heist", "Tremors" und "Van Helsing".
+			</li>
+		</ul>
+	</span>
+	<span slot="left">
+		<img src="/images/visuals/sky_22-01_hilfecenter_peacock_1360x706.jpg" alt="" />
+	</span>
+</SideBySide>
+<SideBySide primary="left">
+	<span slot="left">
+		<h3 class="gradient_text">discovery+ 1 Jahr geschenkt als Extra für alle Sky Q Kunden</h3>
+		<ul>
+			<li>Der Streaming Dienst für großartige Real-Life Unterhaltung die du lieben wirst</li>
+			<li>Mehr als 10.000 Inhalte aller Discovery Sender jederzeit auf Abruf verfügbar</li>
+			<li>
+				Premium Inhalte wie BBC Dokus, discovery+ Originals, Eurosport HD und viele mehr inklusive
+			</li>
+		</ul>
+		<p>
+			1 Jahr lang gratis für alle Sky Q Kunden. Sky Abo abschließen und Gutschein sichern.
+			Aktivierung der App auf dem Sky Q Receiver mit Anmeldung bei discovery+ mittels QR-Code.
+		</p>
+	</span>
+	<span slot="right">
+		<img src="/images/visuals/sky_22-06_discovery-plus_teaser.webp" alt="" />
+	</span>
+</SideBySide>
 
 <style lang="scss">
 	@import 'lib/style/gradient_text.scss';
@@ -36,27 +88,6 @@
 		@include gradient-text;
 	}
 
-	.left {
-		max-width: 800px;
-	}
-	.top_margin {
-		margin-top: 70px;
-	}
-	.side_alignment {
-		display: flex;
-		line-height: 200%;
-		max-width: 85vw;
-		margin-left: auto;
-		margin-right: auto;
-		flex-wrap: wrap;
-		justify-content: center;
-		margin-bottom: 50px;
-	}
-	img {
-		width: 50%;
-		max-width: 500px;
-		object-fit: contain;
-	}
 	h2 {
 		text-align: center;
 		font-size: 38px;
@@ -66,5 +97,9 @@
 		margin-left: auto;
 		margin-right: auto;
 		margin-top: 50px;
+	}
+
+	img {
+		width: 100%;
 	}
 </style>
