@@ -1,9 +1,5 @@
 <script>
 	import PackageTemplate from '$lib/components/templates/element/package_template.svelte';
-
-	import { get_price_string } from '$lib/scripts/universal/asset_library/prices';
-	import GradientBadge from '../../../elements/gradient_badge.svelte';
-	export let detailed = false;
 </script>
 
 <PackageTemplate
@@ -12,18 +8,10 @@
 		'Exklusive Top-Filme kurz nach dem Kino',
 		'Große Filmvielfalt über alle Genres hinweg',
 		'Alle Filme ohne Werbeunterbrechung',
-		'#Sky Q Receiver mit 1TB Festplatte, Ultra HD fähig und Sky On Demand Zugang',
-		'#Sky Go, Sky Kids App - einfach jederzeit auf tausende Titel zugreifen!',
-		'#UHD optional für € 5 mtl.'
+		'Sky Q Receiver mit 1TB Festplatte, Ultra HD fähig und Sky On Demand Zugang',
+		'Sky Go, Sky Kids App - einfach jederzeit auf tausende Titel zugreifen!',
+		'UHD optional für € 5 mtl.'
 	]}
 	image={'/images/assets/packages/normal/cinema.jpg'}
-	{detailed}
->
-	<h3>12 Monate nur {get_price_string(['entertainmentplus', 'cinema'], 'jahr')} mtl.*</h3>
-	<p>
-		(im Jahres-Abo, danach {get_price_string(['entertainmentplus', 'cinema'], 'monat')} mtl. im Monats-Abo)
-		<br />
-		Der Vertrag hat eine Laufzeit von 12 Monaten und ist im Anschluss monatlich kündbar
-	</p>
-	<GradientBadge>50% Rabatt auf Premiumpakete</GradientBadge>
-</PackageTemplate>
+	price_asset_ids={['entertainmentplus', 'cinema']}
+/>
