@@ -1,7 +1,5 @@
 <script>
-	import { get_price_string } from '$lib/scripts/universal/asset_library/prices';
 	import PackageTemplate from '$lib/components/templates/element/package_template.svelte';
-	export let detailed = false;
 </script>
 
 <PackageTemplate
@@ -13,12 +11,5 @@
 		'Einfach. Alles mit einer Sprachfernbedienung'
 	]}
 	image={'/images/assets/packages/normal/entertainmentplus.jpg'}
-	{detailed}
->
-	<h3>12 Monate nur {get_price_string(['entertainmentplus'], 'jahr')} mtl.*</h3>
-	<p>
-		(im Jahres-Abo, danach {get_price_string(['entertainmentplus'], 'monat')} mtl. im Monats-Abo)
-		<br />
-		Der Vertrag hat eine Laufzeit von 12 Monaten und ist im Anschluss monatlich kündbar
-	</p>
-</PackageTemplate>
+	price_asset_ids={['entertainmentplus']}
+/>
