@@ -1,38 +1,39 @@
 <script>
 	import Senders from '$lib/components/layout/senders.svelte';
+	const start = '/images/assets/packages/senders/entertainment/';
 </script>
 
 <body>
 	<Senders
 		text="Sender im Entertainment Paket:"
 		content={[
-			'/images/assets/packages/senders/entertainment/116_sky_logo_13thstreethd_w.png',
-			'/images/assets/packages/senders/entertainment/117_sky_logo_entertainment_w.png',
-			'/images/assets/packages/senders/entertainment/16_sky_logo_romancetv_w.png',
-			'/images/assets/packages/senders/entertainment/388_sky_logo_spiegelgeschichtehd_w.png',
-			'/images/assets/packages/senders/entertainment/452_sky_logo_jukebox.png',
-			'/images/assets/packages/senders/entertainment/52_sky_logo_motorvision_w.png',
-			'/images/assets/packages/senders/entertainment/72_sky_logo_heimatkanal.png',
-			'/images/assets/packages/senders/entertainment/81_sky_logo_natgeochannelhd_w.png',
-			'/images/assets/packages/senders/entertainment/83_sky_logo_discovery-hd_w.png',
-			'/images/assets/packages/senders/entertainment/86_sky_logo_historyhd_w.png',
-			'/images/assets/packages/senders/entertainment/94_sky_logo_boomerang.png',
-			'/images/assets/packages/senders/entertainment/95_sky_logo_cartoonnetwork_w.png',
-			'/images/assets/packages/senders/entertainment/sky_logo_beateuhse-hd.png',
-			'/images/assets/packages/senders/entertainment/sky_logo_crime-investigation_w.png',
-			'/images/assets/packages/senders/entertainment/sky_logo_natgeowild-hd_w.png',
-			'/images/assets/packages/senders/entertainment/sky_logo_sky-atlantic_w.png',
-			'/images/assets/packages/senders/entertainment/sky_logo_sky-comedy_w.png',
-			'/images/assets/packages/senders/entertainment/sky_logo_sky-crime_w.png',
-			'/images/assets/packages/senders/entertainment/sky_logo_sky-documentary_b.png',
-			'/images/assets/packages/senders/entertainment/sky_logo_sky-krimi_w.png',
-			'/images/assets/packages/senders/entertainment/sky_logo_sky-nature_b.png',
-			'/images/assets/packages/senders/entertainment/sky_logo_sky-one_w.png',
-			'/images/assets/packages/senders/entertainment/sky_logo_spiegel-wissen.png',
-			'/images/assets/packages/senders/entertainment/sky_logo_syfyhd_w.png',
-			'/images/assets/packages/senders/entertainment/sky_logo_universalchannelhd_b.png',
-			'/images/assets/packages/senders/entertainment/sky_logo_wb-comedy-hd_b.png',
-			'/images/assets/packages/senders/entertainment/sky_logo_wb-serie-hd_b.png'
-		]}
+			'116_sky_logo_13thstreethd_w.png',
+			//'117_sky_logo_entertainment_w.png',
+			//'16_sky_logo_romancetv_w.png',
+			'388_sky_logo_spiegelgeschichtehd_w.png',
+			//'452_sky_logo_jukebox.png',
+			//'52_sky_logo_motorvision_w.png',
+			//'72_sky_logo_heimatkanal.png',
+			'81_sky_logo_natgeochannelhd_w.png',
+			'83_sky_logo_discovery-hd_w.png',
+			'86_sky_logo_historyhd_w.png',
+			//'94_sky_logo_boomerang.png',
+			//'95_sky_logo_cartoonnetwork_w.png',
+			'sky_logo_beateuhse-hd.png',
+			//'sky_logo_crime-investigation_w.png',
+			'sky_logo_natgeowild-hd_w.png',
+			'sky_logo_sky-atlantic_w.png',
+			'sky_logo_sky-comedy_w.png',
+			'sky_logo_sky-crime_w.png',
+			'sky_logo_sky-documentary_b.png',
+			'sky_logo_sky-krimi_w.png',
+			'sky_logo_sky-nature_b.png',
+			'sky_logo_sky-one_w.png',
+			//'sky_logo_spiegel-wissen.png',
+			'sky_logo_syfyhd_w.png',
+			'sky_logo_universalchannelhd_b.png',
+			'sky_logo_wb-comedy-hd_b.png',
+			'sky_logo_wb-serie-hd_b.png'
+		].map((img) => start + img)}
 	/>
 </body>
