@@ -2,7 +2,7 @@
 	import Button from '../elements/interactive/buttons/button.svelte';
 </script>
 
-<a href="/angebote">
+<a href="/angebote" sveltekit:prefetch>
 	<div class="alignment">
 		<h2>Schnell und einfach Sky abonnieren!</h2>
 		<p>
@@ -10,7 +10,7 @@
 			TAGARO genau richtig. Wir vermitteln Ihnen als autorisierter Sky-Onlinehändler Sky-Verträge
 			bereits seit 2006. Viele Angebote gibt es exklusiv und mit Abschlussprämie.
 		</p>
-		<Button text={'Angebote & Pakete'} route={'/angebote'} />
+		<Button text={'Angebote & Pakete'} on_click={'/angebote'} />
 	</div>
 </a>
 
