@@ -3,7 +3,14 @@ import { indexed_priceable_assets } from './priceable_asset';
 import type { Price } from './priceable_asset_types';
 import type { priceable_asset_id } from './asset_types';
 import { map_entries, typed_entries, typed_from_entries } from 'functional-utilities';
-import { cloneDeep, intersection as intersect, minBy, sum } from 'lodash-es';
+import {
+	cloneDeep,
+	intersection as intersect,
+	isEqual,
+	minBy,
+	sortBy,
+	sum
+} from 'lodash-es';
 import { empty_offer, indexed_offers, offer_applicable, offer_ids } from './offer_description';
 import type { offer_id, offer_description_type } from './offer_description';
 import { asset_sets } from './sets';
@@ -19,7 +26,7 @@ function to_price_string(v: number): string {
 export const aktivierung = 0;
 export const aktivierung_string = to_price_string(aktivierung);
 
-export const bonus = 20;
+export const bonus = 20 as number;
 export const bonus_string = to_price_string(bonus);
 
 const price_table = map_entries(indexed_priceable_assets, ([key, value]) => [key, value.price]);
@@ -29,6 +36,16 @@ export function get_offer_price(
 	assets: ReadonlyArray<priceable_asset_id>
 ): Price | undefined {
 	const current_price_table = cloneDeep(price_table);
+
+	for (const overwrite of offer.overwrites) {
+		if (isEqual(sortBy(overwrite[0]), sortBy(assets))) {
+			const price = get_offer_price(empty_offer, assets);
+			for (const [timeframe, value] of typed_entries(overwrite[1])) {
+				price[timeframe] = value;
+			}
+			return price;
+		}
+	}
 
 	offer.actions.forEach((action) => {
 		const set = asset_sets[action.set];
@@ -40,7 +57,7 @@ export function get_offer_price(
 			current_price_table[asset_id] = typed_from_entries(
 				typed_entries(current_price_table[asset_id]).map(([key, value]) => [
 					key,
-					key === 'jahr' ? operation(action.value[key], value) : value
+					key === 'jahr' ? operation(action.value[key], value as number) : value
 				])
 			);
 		});
@@ -93,15 +110,29 @@ export function sort_by_price(lst: Array<priceable_asset_id>): Array<priceable_a
 	);
 }
 
+function get_offer_savings_string(
+	offer: offer_description_type,
+	packages: ReadonlyArray<package_id>
+): string {
+	return to_price_string(
+		(get_offer_price(empty_offer, packages).jahr - get_offer_price(offer, packages).jahr) * 12
+	);
+}
+export function get_savings_string(packages: ReadonlyArray<package_id>): string {
+	const chosen_offer = chose_offer(packages);
+	if (chosen_offer === undefined) {
+		return to_price_string(0);
+	}
+	return get_offer_savings_string(indexed_offers[chosen_offer], packages);
+}
+
 export function get_offer_note(packages: ReadonlyArray<package_id>, long = false): string {
 	const chosen_offer = chose_offer(packages);
 	if (chosen_offer === undefined) {
 		return '';
 	}
 	const offer = indexed_offers[chosen_offer];
-	if (long) {
-		return offer.long_text;
-	} else {
-		return offer.short_text;
-	}
+	const text = long ? offer.long_text : offer.short_text;
+	return text.replaceAll('{savings}', get_offer_savings_string(offer, packages));
 }
+
