@@ -32,9 +32,12 @@ const adapter = getAdapters(adapt);
 const config = {
 	// Consult https://github.com/sveltejs/svelte-preprocess
 	// for more information about preprocessors
-	preprocess: preprocess({
-		scss: { includePaths: ['src'] }
-	}),
+	preprocess: [
+		preprocess({
+			scss: { includePaths: ['src'] },
+			postcss: true
+		})
+	],
 	kit: {
 		// hydrate the <div id="svelte"> element in src/app.html
 		prerender: {
